Rename QuickSearch folder for consistency
