+++ src/components/Header/components/QuickSearch/PeopleSearch.module.scss
